NameAPI/profile-project-alfie/requirements.txt
Django==3.2.9
djangorestframework
mysqlclient
requests
flake8
autopep8
Pillow
django-cors-headers